pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.46.0",
    "components>=0.0.1a0",
    "newsapi-python>=0.2.7",
    "numpy>=2.2.3",
    "openai>=1.63.2",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "streamlit>=1.42.1",
    "trafilatura>=2.0.0",
    "twilio>=9.4.5",
    "yfinance>=0.2.54",
]
